fix: collect data 구조 변경
- entity class 별도 파일로 분리
- collect_data에 관련 코드 추가
