+++ src/collect_data.py
@@ -4,6 +4,8 @@ import mysql.connector
 from dotenv import load_dotenv
 import os
 
+from model import ParkingLot
+
 load_dotenv()
 
 # mysql-connection 정보
@@ -15,66 +17,15 @@ config = {
     'database':'menudb'
 }
 
-class ParkingLot:
-    def __init__(self, id: int ,reg_id: str, name: str, lat: str, lng: str, sido: str, sigungu: str, full_addr: str, space_no: int):
-        self.__id = id
-        self.__reg_id = reg_id
-        self.__name = name
-        self.__lat = lat
-        self.__lng = lng
-        self.__sido = sido
-        self.__sigungu = sigungu
-        self.__full_addr = full_addr
-        self.__space_no = int(space_no)
-
-    @property
-    def id(self):
-        return self.__id
-
-    @property
-    def reg_id(self):
-        return self.__reg_id
-
-    @property
-    def name(self):
-        return self.__name
-
-    @property
-    def lat(self):
-        return self.__lat
-
-    @property
-    def lng(self):
-        return self.__lng
-
-    @property
-    def sido(self):
-        return self.__sido
-
-    @property
-    def sigungu(self):
-        return self.__sigungu
-
-    @property
-    def full_addr(self):
-        return self.__full_addr
-
-    @property
-    def space_no(self):
-        return self.__space_no
-
-    def __repr__(self):
-        return f'ParkingLot:{self.__id}, {self.__reg_id}, {self.__name}, {self.__lat}, {self.__lng}, {self.__sido}, {self.__sigungu}, {self.__full_addr}'
-
-
 API_KEY = os.getenv("API_KEY")
 BASE_URL = 'https://apis.data.go.kr/B553881/Parking/PrkSttusInfo'
-
-params = {'serviceKey':API_KEY, 'pageNo':1,'numOfRows':500, 'format':2}
 headers = {
     'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/119.0.0.0 Safari/537.36',
     'Accept': 'application/json, text/javascript, */*; q=0.01'
 }
+
+# for pageno in range(1, PAGE_NO):
+params = {'serviceKey':API_KEY, 'pageNo':1,'numOfRows':2000, 'format':2}
 response = requests.get(url=BASE_URL, params=params, headers=headers)
 
 parking_lots_list: list[ParkingLot] = []
@@ -125,3 +76,42 @@ if len(parking_lots_list) > 0:
                 conn.commit()
     except mysql.connector.Error as err:
         print('DB에러: ', err)
+
+# executemany 참고 코드
+"""
+if len(parking_lots_list) > 0:
+    try:
+        with mysql.connector.connect(**config) as conn:
+            with conn.cursor() as cursor:
+
+                # %s 순서: reg_id, name, lat, lng, sido, sigungu, addr, space, lng, lat
+                sql = '''
+                    INSERT INTO parkinglot (
+                        reg_id, name, lat, lng, sido, sigungu, full_address, space_no, coord
+                    ) VALUES (
+                        %s, %s, %s, %s, %s, %s, %s, %s, 
+                        ST_GeomFromText(CONCAT('POINT(', %s, ' ', %s, ')'), 4326, 'axis-order=long-lat')
+                    )
+                '''
+
+                # 2. 데이터 리스트 빌드 (List of Tuples)
+                data_list = [
+                    (
+                        p.reg_id, p.name, p.lat, p.lng, p.sido, p.sigungu, p.full_addr, p.space_no,
+                        p.lng, # POINT 경도
+                        p.lat  # POINT 위도
+                    ) for p in parking_lots_list
+                ]
+
+                # 3. 일괄 삽입 실행
+                cursor.executemany(sql, data_list)
+                
+                conn.commit()
+                print(f"총 {cursor.rowcount}개의 데이터가 성공적으로 삽입되었습니다.")
+
+    except mysql.connector.Error as err:
+        print('DB에러: ', err)
+
+"""
+
+#
